Separated code into multiple classes.
